requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
pyrogram==2.0.106
tgcrypto==1.2.5
aiofiles==23.2.1
aiohttp==3.9.1
pydantic-settings==2.1.0
python-multipart==0.0.6
supabase==2.0.2
psycopg2-binary==2.9.9
asyncpg==0.29.0
pillow==10.1.0
python-magic==0.4.27
tenacity==8.2.3
